Conversational_UI/Releases/NLPEngine/requirements.txt
kafka-python~=2.0.2
nltk~=3.5
scikit-learn~=0.24.1
flask~=1.1.2